% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{overDueRpt_Query}
\alias{overDueRpt_Query}
\title{业务应收逾期报表查询}
\usage{
overDueRpt_Query(
  erp_token = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FDate = "2022-05-12",
  Fdepartment = "方案中心",
  FBusinessGroup = "华东区",
  FBusinessMan = "hawken",
  FCompanyScope = "所有公司",
  FDataScope = "本部门"
)
}
\arguments{
\item{erp_token}{口令}

\item{FDate}{日期}

\item{Fdepartment}{部门}

\item{FBusinessGroup}{业务组}

\item{FBusinessMan}{业务员}

\item{FCompanyScope}{公司范围}

\item{FDataScope}{数据范围}
}
\value{
返回值
}
\description{
业务应收逾期报表查询
}
\examples{
overDueRpt_Query()
}
