% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdueRpt.R
\name{overdueRptServer}
\alias{overdueRptServer}
\title{业务应收逾期报表}
\usage{
overdueRptServer(input, output, session, app_id)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{app_id}{程序ID}
}
\value{
返回值
}
\description{
业务应收逾期报表
}
\examples{
overdueRptServer()
}
